requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cvzone==1.6.1
distro==1.9.0
Eel==0.18.2
Flask==3.1.1
Flask-MySQLdb==2.0.0
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hugchat==0.5.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
narwhals==1.47.0
numpy==2.3.1
openai==1.95.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==11.2.1
playsound==1.2.2
protobuf==6.31.1
pvporcupine==3.0.5
pyarrow==20.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==310
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wikipedia==1.4.0
zope.event==5.1
zope.interface==7.2
